remove image from storage when delete from show view

+++ app/Http/Controllers/Admin/ExamController.php
@@ -281,7 +281,7 @@ class ExamController extends Controller
         ]);
 
         $image_path = '';
-        if ($request->hasFile('img')) {
+        if ($request->img !== null && $request->img !== '') {
             $image_path = $request->file('img')->store('img', 'public');
             File::delete(storage_path('app/public/' . $request->img2));
             // $del_image = Question::find($image_path);
@@ -322,7 +322,7 @@ class ExamController extends Controller
             'answer'            => 'required',
         ]);
 
-        if ($request->condition == 3) {
+        if ($request->condition == 3 && ($request->img2 !== null && $request->img2 !== '')) {
             File::delete(storage_path('app/public/' . $request->img2));
             // $del_image = Question::find($image_path);
             // File::delete($del_image);
